feat: Improve dependent field with same struct
