+++ src/models/process.rs
@@ -3,7 +3,7 @@ use salvo::prelude::*;
 use serde_json::{json, Value};
 use std::{io::ErrorKind, io::Error};
 
-use crate::{controller::opensearch::{IndexDocument, find, get_by_id}, models::get_index_name};
+use crate::{controller::opensearch::{IndexDocument, find, get_by_id}, models::{generic::DependendField, get_index_name}};
 
 #[derive(Deserialize, Extractible, Debug, Clone)]
 #[salvo(extract(default_source(from = "body")))]
@@ -27,17 +27,6 @@ impl Default for ProcessResponse {
             process: None
         }
     }
-}
-
-#[derive(Deserialize, Serialize, Extractible, Debug, Clone)]
-pub struct DependendField {
-    pub uuid: Option<String>,
-    pub internal_id: Option<i32>,
-    pub id: Option<String>,
-    pub column_name: Option<String>,
-    pub parent_id: Option<i32>,
-    pub parent_uuid: Option<String>,
-    pub parent_name: Option<String>
 }
 
 #[derive(Deserialize, Serialize, Extractible, Debug, Clone)]
